Part 1: Re-organize project by features
  - Create src/features and src/shared directories
  - Move TodoForm.jsx to src/features/
  - Move TodoList.jsx and TodoListItem.jsx to src/features/TodoList/
  - Update import statements in App.jsx
  - Project now organized by feature modules
