+++ src/App.jsx
@@ -3,8 +3,8 @@ import { useState } from 'react'
 import reactLogo from './assets/react.svg'
 import viteLogo from '/vite.svg'
 import './App.css'
-import TodoForm from './TodoForm'
-import TodoList from './TodoList'
+import TodoForm from './features/TodoForm'
+import TodoList from './features/TodoList/TodoList'
 
 function App() {
     const [count, setCount] = useState(0)
